Move user choice method to new file
