+++ app.rb
@@ -5,6 +5,7 @@ require_relative './teacher'
 require_relative './student'
 require_relative './rental'
 require 'date'
+require './menu'
 
 # app class
 class Apps
@@ -135,27 +136,5 @@ class Apps
   # convert date
   def convert_date(str)
     Date.parse(str)
-  end
-
-  def user_choice(choice)
-    case choice
-    when 1
-      list_books
-    when 2
-      list_people
-    when 3
-      create_person
-    when 4
-      create_book
-    when 5
-      create_rental
-    when 6
-      list_rentals
-    when 7
-      puts 'Thank you for using the school library app'
-      exit
-    else
-      puts ' Choose a number between 1 and 7'
-    end
   end
 end
